refactor: useSpaceDisplay -> useSpacesDisplay 로 수정
