+++ frontend/src/hooks/domain/useSpacesDisplay.ts
