+++ frontend/src/pages/host/mypage/MyPage.tsx
@@ -5,7 +5,7 @@ import Dropdown, {
 } from '../../../components/@common/dropdown/Dropdown';
 import Thumbnail from '../../../components/@common/thumbnail/Thumbnail';
 import SpaceCard from '../../../components/host/spaceCard/SpaceCard';
-import useSpacesDisplay from '../../../hooks/domain/useSpaceDisplay';
+import useSpacesDisplay from '../../../hooks/domain/useSpacesDisplay';
 import { MyPageMockData, SpaceMockData } from '../../mockData';
 import * as S from './MyPage.styles';
